requirements.txt
Flask==1.0.2
line-bot-sdk==1.8.0
urllib3==1.23
requests==2.19.1
pillow
keras
opencv-python
tensorflow
